requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4
Pillow==10.1.0
pyTelegramBotAPI==4.14.0
requests==2.31.0
urllib3==2.0.7
